wolfssl: reorganized build options
Removed options that can't be turned off because we're building
with --enable-stunnel, some of which affected hostpad's Config.in.
Adjusted the title of OCSP option, as OCSP itself can't be turned off,
only the stapling part is selectable.
Fixed poly1035 build.
Mark options turned on when wpad support is selected.
Add building with TLS v1.0 support as an option.

Signed-off-by: Eneas U de Queiroz <cote2004-github@yahoo.com>