libtool: bump to 2.4.6
This updated libtool to its current release, from 2015.  It has changed
the way bootstrap gets libgnu code.  It ordinarily fetches it using git.
We are not doing it, since we apply patches to its sources.

Current patches were renumbered and given a description text.  The fix
in 160-passthrough-ssp.patch is no longer needed.

A patch to speed up build was cherry-picked, and another openwrt
specific patch was needed to not use quotes in $(SHELL), to acommodate
our "SHELL=/usr/bin/env bash" usage.

The already present call to ./bootstrap ensures that generated files are
refreshed, so the patches are applied only to their sources.  Also, that
bootstrap call was adjusted to run at the appropriate time when QUILT=1.

Signed-off-by: Eneas U de Queiroz <cotequeiroz@gmail.com>