mac80211: always allow pre-CAC
Function regulatory_pre_cac_allowed only allows pre-CAC only if the
regdomain is ETSI, even though the comments in cfg80211.h mentions that
it is allowed only in some regdomains (notable ETSI).
Notable is not a synonim for 'only'. It turns out that, at least in
Brazil, there's no provision disallowing it.

Since this is used privately here in Brazil, I'll take the easy way and just
always return true.

Signed-off-by: Eneas U de Queiroz <cotequeiroz@gmail.com>