build: kernel: avoid unnecessary package rebuilds
The $(LINUX_DIR)/.config changes between runs of
make target/compile (which only compiles modules) and
make target/install (which compiles the image).

The kerenl packages are build in between those runs, and they check the
.config timestamp to decide if they need to be rebuilt.

Save the target/compile .config to use its timestamp if the file does
not change between runs.  That way the subsequent kernel packages are
not unnecessarily rebuilt.

Signed-off-by: Eneas U de Queiroz <cotequeiroz@gmail.com>