hostapd: avoid unnecessary package rebuilds
Package hostapd-common is currently built for every different hostapd
variant.  This is not only unnecessary, but causes a rebuild of every
variant selected if you compile it back to back.

Avoid that by using its own variant to build the hostapd-common package,
and move it to the top of the build list, so that subsequent variantes
will not try to build it over and over again.

Signed-off-by: Eneas U de Queiroz <cotequeiroz@gmail.com>