mac80211: avoid unnecessary rebuilds
Move the order in which BuildPackage is called, so that the driver
packages are built after cfg80211 and mac80211, to avoid forcing a
clean-build of the package when 'make package/mac80211/compile' is
called a second time without changes.

Because both mac80211 and cfg80211 have the ABI_VERSION field, dependent
pacakges will check mac80211 and cfg80211 .version files; their
timestamp should be older than the dependent package's STAMP_PREPARED
file.

Signed-off-by: Eneas U de Queiroz <cotequeiroz@gmail.com>